+++ public/components/BottomBar/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+import BottomBar from "./BottomBar";
+
+export default BottomBar;
